Video - 26-12/requirements.txt
fastapi
uvicorn
fpdf
nltk
moviepy
ffmpeg-python
google-api-python-client
google-cloud
google-cloud-vision
openai-whisper==20230918
